my_app/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
contourpy==1.0.6
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.16
django-bootstrap4==22.3
django-crispy-forms==1.14.0
django-import-export==3.0.1
django-rangefilter==0.9.1
django-relativedelta==1.1.2
et-xmlfile==1.1.0
fonttools==4.38.0
gunicorn==20.1.0
kiwisolver==1.4.4
MarkupPy==1.14
matplotlib==3.6.2
numpy==1.23.5
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.7
xlrd==2.0.1
xlwt==1.3.0
